Moved the interfaces to types
