--- src/interfaces/interfaces.tsx
@@ -1,21 +1,0 @@
-export interface User {
-  contacts: ContactData[]
-  visits: VisitData[]
-}
-
-export interface VisitData {
-  displayName: string
-  petNames: string[]
-  dateTime: string
-  duration: string
-  walkDist: number
-  commuteDist: number
-  commuteMethod: string
-  notes: string
-}
-
-export interface ContactData {
-  displayName: string
-  email: string
-  contactNumber: string
-}
