+++ src/types/types.ts
@@ -1,12 +1,28 @@
-interface Pet {
+export interface Pet {
   id: string
   name: string
   notes: string
 }
-interface Contact {
+
+export interface User {
+  contacts: ContactData[]
+  visits: VisitData[]
+}
+
+export interface VisitData {
+  displayName: string
+  petNames: string[]
+  dateTime: string
+  duration: string
+  walkDist: number
+  commuteDist: number
+  commuteMethod: string
+  notes: string
+}
+
+export interface ContactData {
   id: string
-  firstName: string
-  lastName: string
+  displayName: string
   desc: string
   pets: Pet[]
   email: string
@@ -16,5 +32,3 @@ interface Contact {
   notes: string
   tags: string[]
 }
-
-export type { Contact, Pet }
